refactor: centralize markers
